+++ src/routes/(v2)/v2/bigint/schema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+	number: z.bigint().min(BigInt(1000000), 'A BIG number please!')
+});